Improve SearchLocationView
I'm really happy with this commit --> the SearchLocationView works just as I want it to, giving the user the
ability to select any city/place in the world and set their location to that spot. This way, they can search for
food anywhere they'd like. I also refactored some of the code in FilterView so that it looks better and is easier
to read. Next, I want to either make sure the map stays in the same spot where the user left off or put it to
the user's current location